requirements.txt
ruff==0.11.5
python-telegram-bot==22.0
eth-brownie==1.21.0
# `brownie-safe==0.9.0` has conflicting dependencies with `eth-brownie==1.21.0`, so using wavey's fork
git+https://github.com/wavey0x/brownie-safe.git@deps
setuptools==80.9.0